// app/models/IKpis.ts
export type KpiResponse = {
  totalBuffets: number;
  totalArchived: number;
  totalEnquiries: number;
  pendingEnquiries: number;
  resolvedEnquiries: number;
  totalReservations: number;
  reservationsMadeLastWeek: number;
  reservationsMadeThisWeek: number;
  totalPaymentsMade: number;
  totalPaymentsMadeLastWeek:number;
  totalPaymentsMadeThisWeek: number;
};